+++ docs/public/release_notes.rst
@@ -27,6 +27,8 @@ New features:
 
 Fixes:
 
+- Ecountering a regular (un-translatable) model in a deep `select_related` does
+  not break anymore. — :issue:`206`.
 - Language tabs URI are now correctly generated when changelist filters are used.
   — :issue:`203`.
 - Admin language tab selection is no longer lost when change filters are active.